# Removes resource consuming Redis subscription process from a global lock