Removes resource consuming Redis subscription process from a global lock
- Adds a new structure to hold connection count to each subscription
- Makes global lock much shorter