@Boychenko 2018-05-17
Could you point me what tests did you run?