@Boychenko 2018-05-17
@BrennanConroy It will be much better if you give the command to execute to see perf results