@anurse 2018-05-17
The initial change to use this locking was to solve a correctness issue, we only did ad-hoc performance testing to verify that it didn't significantly degrade performance. In order to take a PR refactoring this code in the name of performance we need to see data showing the performance has improved. We don't currently have pre-written benchmarks for this scenario, so we don't have a specific command to run.

What prompted this change? Are you seeing a performance impact in a specific scenario? We generally avoid performance optimizations unless there's specific data (memory profiles, traces, benchmarks) driving the decision.