@Boychenko 2018-05-18
@anurse I've made a new benchmark for concurrent connections. To show improvement I've added network delay 1ms for all methods `TestRedisServer`.
Result:
### Before
``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.431)
Intel Core i7-4710HQ CPU 2.50GHz (Haswell), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2435776 Hz, Resolution=410.5468 ns, Timer=TSC
.NET Core SDK=2.1.300-rc1-008673
  [Host]     : .NET Core 2.1.0-rtm-26509-04 (CoreCLR 4.6.26509.03, CoreFX 4.6.26509.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-rtm-26509-04 (CoreCLR 4.6.26509.03, CoreFX 4.6.26509.03), 64bit RyuJIT


```
|    Method | UserCount | GroupCount |      Mean |     Error |    StdDev |   Op/s |   Gen 0 | Allocated |
|---------- |---------- |----------- |----------:|----------:|----------:|-------:|--------:|----------:|
| **JoinLeave** |         **1** |          **1** |  **15.49 ms** | **0.6007 ms** | **0.5619 ms** | **64.545** | **31.2500** |    **7.8 KB** |
| **JoinLeave** |         **1** |         **15** |  **69.25 ms** | **0.3711 ms** | **0.3471 ms** | **14.441** |       **-** |    **7.8 KB** |
| **JoinLeave** |        **20** |          **1** |  **80.20 ms** | **0.4067 ms** | **0.3606 ms** | **12.470** | **62.5000** |    **7.8 KB** |
| **JoinLeave** |        **20** |         **15** | **109.05 ms** | **2.1173 ms** | **2.3534 ms** |  **9.170** | **62.5000** |    **7.8 KB** |

### After
``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.431)
Intel Core i7-4710HQ CPU 2.50GHz (Haswell), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2435776 Hz, Resolution=410.5468 ns, Timer=TSC
.NET Core SDK=2.1.300-rc1-008673
  [Host]     : .NET Core 2.1.0-rtm-26509-04 (CoreCLR 4.6.26509.03, CoreFX 4.6.26509.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.0-rtm-26509-04 (CoreCLR 4.6.26509.03, CoreFX 4.6.26509.03), 64bit RyuJIT


```
|    Method | UserCount | GroupCount |     Mean |     Error |    StdDev |   Median |  Op/s |   Gen 0 | Allocated |
|---------- |---------- |----------- |---------:|----------:|----------:|---------:|------:|--------:|----------:|
| **JoinLeave** |         **1** |          **1** | **13.66 ms** | **0.2730 ms** | **0.5260 ms** | **13.48 ms** | **73.19** | **31.2500** |    **7.8 KB** |
| **JoinLeave** |         **1** |         **15** | **12.27 ms** | **0.2436 ms** | **0.4002 ms** | **12.19 ms** | **81.49** | **31.2500** |    **7.8 KB** |
| **JoinLeave** |        **20** |          **1** | **20.87 ms** | **0.2816 ms** | **0.2635 ms** | **20.88 ms** | **47.92** | **31.2500** |    **7.8 KB** |
| **JoinLeave** |        **20** |         **15** | **11.97 ms** | **0.3046 ms** | **0.7120 ms** | **11.73 ms** | **83.53** | **62.5000** |    **7.8 KB** |