@anurse 2018-05-18
Awesome work @Boychenko ! We try to make sure that when we're taking a performance-related PR we know precisely what benefit it has, so thanks very much for the benchmark data. We'll take a look at the data and the PR!