@BrennanConroy 2018-05-18
@Boychenko Nice work writing a benchmark! Unfortunately the `TestRedisServer` is very untested and has bugs around keeping subscriptions around which heavily affects your benchmarking scenarios making the numbers unusable.

I'm curious, did you see a real world slow down because of this lock?

We can't take this change unless we know for a fact that there is real perf concerns.