@Boychenko 2018-05-18
@BrennanConroy What bugs in `TestRedisServer` affected my benchmarking scenarios?
I think real world scenario is quite obvious. When a few users connect to the server or join groups at the same time global lock doesn't make this process faster.
I'm not saying that library starts to be not usable without this improvement but it can be faster.