@anurse 2018-08-08
I'm going to close this. I think we should gather more data on this if we want to take a performance improvement here. At the moment, I don't think this is a high priority, but if we start getting feedback about performance problems here we can revisit it.